[refactoring] code(fix, norm)
