+++ Libft/ft_strlen.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jinsecho <jinsecho@student.42gyeongsa      +#+  +:+       +#+        */
+/*   By: jinsecho <jinsecho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/26 20:17:11 by jinsecho          #+#    #+#             */
-/*   Updated: 2024/02/29 17:55:41 by jinsecho         ###   ########.fr       */
+/*   Updated: 2024/10/28 15:51:16 by jinsecho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,8 @@ int	ft_strlen(const char *string)
 	int	len;
 
 	len = 0;
+	if (string == NULL)
+		return (len);
 	while (*string)
 	{
 		string++;
