+++ main.c
@@ -48,146 +48,6 @@ int	min_exit_num_check(char num, int i)
 	return (0);
 }
 
-static void	replace_env_find(t_cmd *cmd, char **ptr, char *line, int *n)
-{
-	char	*tmp;
-	int		i;
-
-	i = 0;
-	while (ft_isalnum(line[*n]) || line[*n] == '_')
-		(*n)++;
-	tmp = (char *)malloc(sizeof(char) * (*n));
-	if (tmp == NULL)
-		exit (EXIT_FAILURE);
-	ft_strlcpy(tmp, line + 1, *n);
-	while (cmd->envp[i])
-	{
-		if (ft_envchr(cmd->envp[i], tmp))
-		{
-			*ptr = ft_strdup(ft_strchr(cmd->envp[i], '=') + 1);
-			free(tmp);
-			return ;
-		}
-		i++;
-	}
-	*ptr = ft_strdup("");
-	free(tmp);
-}
-
-static char	*ms_str_replace_env(t_cmd *cmd, char *line, int *n)
-{
-	char	*ptr;
-
-	*n = 1;
-	if (line[*n] == '?')
-	{
-		*n = 2;
-		ptr = ft_itoa(0);
-	}
-	else if (ft_isalpha(line[*n]) || line[*n] == '_')
-		replace_env_find(cmd, &ptr, line, n);
-	else if (ft_isdigit(line[*n]))
-	{
-		*n = 2;
-		ptr = ft_strdup("");
-	}
-	else
-		ptr = ft_strdup("");
-	return (ptr);
-}
-
-void	ms_line_replace_env(t_cmd *cmd, char **ptr, char *line)
-{
-	t_env_var	env;
-
-	env.cnt = 0;
-	while ((*ptr)[env.cnt] != '$' && (*ptr)[env.cnt])
-		(env.cnt)++;
-	while ((*ptr)[env.cnt])
-	{
-		env.env_tmp = ms_str_replace_env(cmd, &(*ptr)[env.cnt], &env.n);
-		if (env.n == 1)
-			(env.cnt)++;
-		else
-		{
-			env.len = ft_strlen(line) + ft_strlen(env.env_tmp);
-			*ptr = ft_realloc(*ptr, env.len + 1);
-			env.str_tmp = ft_strdup(&(*ptr)[env.cnt] + env.n);
-			ft_strlcpy(&(*ptr)[env.cnt], env.env_tmp, ft_strlen(env.env_tmp) + 1);
-			ft_strlcat(*ptr, env.str_tmp, env.len + 1);
-			(env.cnt) += ft_strlen(env.env_tmp);
-			free(env.str_tmp);
-		}
-		while ((*ptr)[env.cnt] != '$' && (*ptr)[env.cnt])
-			(env.cnt)++;
-		free(env.env_tmp);
-	}
-}
-
-int	quote_flag_len(char *line, char c, int *i, int *flag)
-{
-	int	k;
-	
-	k = 0;
-	if (c == DOUBLE_QUOTE)
-		*flag = 1;
-	while (line[*i + k] != c && line[*i + k])
-		k++;
-	return (k);
-}
-
-char	*ms_line_tokenizing_quote(t_cmd *cmd, char *line, int *i)
-{
-	char	*ptr;
-	int		k;
-	int		quote_flag;
-
-	quote_flag = 0;
-	ptr = (char *)malloc(sizeof(char) * ft_strlen(line) + 1);
-	if (ptr == NULL)
-		exit (EXIT_FAILURE);
-	if (line[(*i)++] == DOUBLE_QUOTE)
-		k = quote_flag_len(line, DOUBLE_QUOTE, i, &quote_flag);
-	else
-		k = quote_flag_len(line, SINGLE_QUOTE, i, &quote_flag);
-	ft_strlcpy(ptr, line + *i, k + 1);
-	(*i) += k;
-	if (line[*i] == 0)
-	{
-		free(ptr);
-		return (NULL);
-	}
-	else
-		(*i)++;
-	if (quote_flag)
-		ms_line_replace_env(cmd, &ptr, line);
-	return (ptr);
-}
-
-char	*ms_line_tokenizing_str(t_cmd *cmd, char *line, int *i)
-{
-	char	*ptr;
-	int		k;
-	
-	k = 0;
-	ptr = (char *)malloc(sizeof(char) * ft_strlen(line) + 1);
-	if (ptr == NULL)
-		exit (EXIT_FAILURE);
-	if (ft_strchr("|><", line[*i]))
-		k++;
-	else
-	{
-		while (ft_strchr(" |><\t\n", line[*i + k]) == NULL \
-		&& line[*i + k] != DOUBLE_QUOTE \
-		&& line[*i + k] != SINGLE_QUOTE && line[*i + k])
-		k++;
-	}
-	ft_strlcpy(ptr, line + *i, k + 1);
-	(*i) += k;
-	ms_line_replace_env(cmd, &ptr, line);
-	return (ptr);
-}
-
 void	ms_line_tokenizer(t_cmd *cmd, char *line)
 {
 	int		i = 0;
